+++ examples/lib/stories/animations/basic.dart
@@ -59,8 +59,7 @@ class BasicAnimations extends BaseGame with TapDetector {
       removeOnFinish: true,
     );
 
-    animationComponent.position = position;
-    animationComponent.position = animationComponent.position - size / 2;
+    animationComponent.position = position - size / 2;
     add(animationComponent);
   }