+++ examples/lib/stories/controls/draggables.dart
@@ -39,10 +39,10 @@ class DraggableSquare extends PositionComponent
       return false;
     }
 
-    final localCoords = gameRef!.convertGlobalToLocalCoordinate(
+    final localCoords = gameRef.convertGlobalToLocalCoordinate(
       details.globalPosition.toVector2(),
     );
-    position = localCoords - dragDeltaPosition;
+    position.setFrom(localCoords - dragDeltaPosition);
     return false;
   }