+++ packages/flame/lib/src/components/joystick/joystick_directional.dart
@@ -50,7 +50,7 @@ class JoystickDirectional extends BaseComponent with Draggable, HasGameRef {
 
   @override
   Future<void> onLoad() async {
-    initialize(gameRef!.size);
+    initialize(gameRef.size);
   }
 
   @override
@@ -163,7 +163,7 @@ class JoystickDirectional extends BaseComponent with Draggable, HasGameRef {
   @override
   bool onDragUpdate(int pointerId, DragUpdateDetails details) {
     if (_dragging) {
-      _dragPosition = gameRef!.convertGlobalToLocalCoordinate(
+      _dragPosition = gameRef.convertGlobalToLocalCoordinate(
         details.globalPosition.toVector2(),
       );
       return false;