+++ packages/flame/lib/src/components/mixins/has_game_ref.dart
@@ -1,5 +1,26 @@
+import '../../../components.dart';
 import '../../game/game.dart';
 
 mixin HasGameRef<T extends Game> {
-  T? gameRef;
+  T? _gameRef;
+
+  T get gameRef {
+    final ref = _gameRef;
+    if (ref == null) {
+      throw 'Accessing gameRef before the component was added to the game!';
+    }
+    return ref;
+  }
+
+  bool get hasGameRef => _gameRef != null;
+
+  set gameRef(T gameRef) {
+    _gameRef = gameRef;
+    if (this is BaseComponent) {
+      (this as BaseComponent)
+          .children
+          .whereType<HasGameRef<T>>()
+          .forEach((e) => e.gameRef = gameRef);
+    }
+  }
 }