+++ packages/flame/lib/src/effects/scale_effect.dart
@@ -39,7 +39,7 @@ class ScaleEffect extends SimplePositionComponentEffect {
   @override
   void initialize(PositionComponent component) {
     super.initialize(component);
-    _startSize = component.size;
+    _startSize = component.size.clone();
     _delta = isRelative ? size : size - _startSize;
     if (!isAlternating) {
       endSize = _startSize + _delta;
@@ -52,6 +52,6 @@ class ScaleEffect extends SimplePositionComponentEffect {
   @override
   void update(double dt) {
     super.update(dt);
-    component?.size = _startSize + _delta * curveProgress;
+    component?.size.setFrom(_startSize + _delta * curveProgress);
   }
 }