+++ packages/flame/test/components/resizable_test.dart
@@ -5,7 +5,7 @@ import 'package:test/test.dart';
 class MyComponent extends PositionComponent {
   String name;
   @override
-  Vector2 size = Vector2(2.0, 2.0);
+  final Vector2 size = Vector2(2.0, 2.0);
   late Vector2 gameSize;
 
   MyComponent(this.name);