+++ packages/flame/test/game/camera_and_viewport_test.dart
@@ -11,10 +11,13 @@ import '../util/mock_canvas.dart';
 
 class TestComponent extends PositionComponent {
   static final Paint _paint = Paint();
-  TestComponent(Vector2 position) {
-    this.position = position;
-    size = Vector2.all(1.0);
-  }
+
+  TestComponent(Vector2 position)
+      : super(
+          position: position,
+          size: Vector2.all(1.0),
+        );
+
   @override
   void render(Canvas c) {
     super.render(c);
@@ -38,7 +41,7 @@ void main() {
       expect(game.size, Vector2.all(50.00));
 
       final viewport = game.viewport as FixedResolutionViewport;
-      expect(viewport.resizeOffset, Vector2(0, 0));
+      expect(viewport.resizeOffset, Vector2.zero());
       expect(viewport.scaledSize, Vector2(200.0, 200.0));
       expect(viewport.scale, 4.0);
 
@@ -146,6 +149,7 @@ void main() {
       game.camera.moveTo(Vector2.all(4.0));
       game.camera.snap();
       // the component will now be at 10 - 4 = (6, 6)
+      expect(game.camera.position, Vector2.all(4.0));
 
       final canvas = MockCanvas();
       game.render(canvas);
@@ -181,10 +185,10 @@ void main() {
       final p = TestComponent(Vector2.all(10.0))..anchor = Anchor.center;
       game.add(p);
       game.update(0);
-      game.camera.followObject(p);
+      game.camera.followComponent(p);
 
       expect(game.camera.position, Vector2.all(0.0));
-      p.position = Vector2(10.0, 20.0);
+      p.position.setValues(10.0, 20.0);
       // follow happens immediately because the object's movement is assumed to be smooth
       game.update(0);
       // (10,20) - half screen (50,50)
@@ -210,10 +214,10 @@ void main() {
       game.add(p);
       game.update(0);
       // this would be a typical vertical shoot-em-up
-      game.camera.followObject(p, relativeOffset: Vector2(0.5, 0.8));
+      game.camera.followComponent(p, relativeOffset: Vector2(0.5, 0.8));
 
       expect(game.camera.position, Vector2.all(0.0));
-      p.position = Vector2(600.0, 2000.0);
+      p.position.setValues(600.0, 2000.0);
       // follow happens immediately because the object's movement is assumed to be smooth
       game.update(0);
       // (600,2000) - fractional screen (50,80)
@@ -238,29 +242,29 @@ void main() {
       final p = TestComponent(Vector2.all(10.0))..anchor = Anchor.center;
       game.add(p);
       game.update(0);
-      game.camera.followObject(
+      game.camera.followComponent(
         p,
         worldBounds: const Rect.fromLTWH(-1000, -1000, 2000, 2000),
       );
 
-      p.position = Vector2(600.0, 700.0); // well within bounds
+      p.position.setValues(600.0, 700.0); // well within bounds
       game.update(0);
       expect(game.camera.position, Vector2(550, 650));
 
       // x ok, y starts to get to bounds
-      p.position = Vector2(600.0, 950.0); // right on the edge
+      p.position.setValues(600.0, 950.0); // right on the edge
       game.update(0);
       expect(game.camera.position, Vector2(550, 900));
 
-      p.position = Vector2(600.0, 950.0); // stop advancing
+      p.position.setValues(600.0, 950.0); // stop advancing
       game.update(0);
       expect(game.camera.position, Vector2(550, 900));
 
-      p.position = Vector2(-1100.0, 950.0);
+      p.position.setValues(-1100.0, 950.0);
       game.update(0);
       expect(game.camera.position, Vector2(-1000, 900));
 
-      p.position = Vector2(1000.0, 1000.0);
+      p.position.setValues(1000.0, 1000.0);
       game.update(0);
       expect(game.camera.position, Vector2(900, 900));
     });
@@ -271,7 +275,7 @@ void main() {
       final p = TestComponent(Vector2.all(10.0))..anchor = Anchor.center;
       game.add(p);
       game.update(0);
-      game.camera.followObject(
+      game.camera.followComponent(
         p,
         worldBounds: const Rect.fromLTWH(0, 0, 100, 100),
       );
@@ -280,13 +284,47 @@ void main() {
       game.update(0);
       expect(game.camera.position, Vector2(50, 50));
 
-      p.position = Vector2(60.0, 50.0);
+      p.position.setValues(60.0, 50.0);
       game.update(0);
       expect(game.camera.position, Vector2(50, 50));
 
-      p.position = Vector2(-10.0, -20.0);
+      p.position.setValues(-10.0, -20.0);
       game.update(0);
       expect(game.camera.position, Vector2(50, 50));
+    });
+    test('camera relative offset without follow', () {
+      final game = BaseGame();
+      game.onResize(Vector2.all(200.0));
+
+      game.camera.setRelativeOffset(Anchor.center.toVector2());
+
+      game.update(0);
+      expect(game.camera.position, Vector2.zero());
+
+      game.update(10000);
+      expect(game.camera.position, Vector2.all(-100.0));
+    });
+    test('camera zoom', () {
+      final game = BaseGame();
+      game.onResize(Vector2.all(200.0));
+      game.camera.zoom = 2;
+
+      final p = TestComponent(Vector2.all(100.0))..anchor = Anchor.center;
+      game.add(p);
+      game.update(0);
+
+      final canvas = MockCanvas();
+      game.render(canvas);
+      expect(
+        canvas.methodCalls
+            .where((e) => e.startsWith('translate') || e.startsWith('scale')),
+        [
+          'translate(0.0, 0.0)', // camera translation
+          'scale(2.0)', // camera zoom
+          'translate(100.0, 100.0)', // position component
+          'translate(-0.5, -0.5)', // anchor
+        ],
+      );
     });
   });
   group('viewport & camera', () {
@@ -299,7 +337,7 @@ void main() {
 
       final p = TestComponent(Vector2.all(10.0))..anchor = Anchor.center;
       game.add(p);
-      game.camera.followObject(
+      game.camera.followComponent(
         p,
         // this could be a typical mario-like platformer, where the player is more on the bottom left to allow the scenario to be seem
         relativeOffset: Vector2(0.25, 0.25),
@@ -309,19 +347,19 @@ void main() {
       game.update(0);
       expect(game.camera.position, Vector2(0, 0));
 
-      p.position = Vector2(30.0, 0.0);
+      p.position.setValues(30.0, 0.0);
       game.update(0);
       expect(game.camera.position, Vector2(5, 0));
 
-      p.position = Vector2(30.0, 100.0);
+      p.position.setValues(30.0, 100.0);
       game.update(0);
       expect(game.camera.position, Vector2(5, 75));
 
-      p.position = Vector2(30.0, 1000.0);
+      p.position.setValues(30.0, 1000.0);
       game.update(0);
       expect(game.camera.position, Vector2(5, 900));
 
-      p.position = Vector2(950.0, 20.0);
+      p.position.setValues(950.0, 20.0);
       game.update(0);
       expect(game.camera.position, Vector2(900, 0));
     });